fix(auth): break server-only import chain from client signin page
auth-helpers.ts imports from ./index.ts (the auth barrel), which imports
@pagespace/db directly. signin/page.tsx ('use client') imported
isSafeNextPath and SIGNIN_NEXT_ALLOWED_PREFIXES from auth-helpers, pulling
pg into the client bundle. pg@8.21.0 introduced require('util/types') — a
subpath specifier webpack cannot polyfill for browsers — exposing this.

- Extract pure URL validation utilities (isSafeReturnUrl, isSafeNextPath,
  SIGNIN_NEXT_ALLOWED_PREFIXES) into url-utils.ts with no server imports
- Add 'server-only' guard to auth-helpers.ts to catch future mistakes
  at build time instead of bundle time
- Point signin/page.tsx at url-utils directly
- Stub server-only in vitest.config.ts alias (jsdom has no react-server
  export condition)
- Remove the pg <8.21.0 override — now that no client component imports pg,
  any pg version works correctly
- Update next.config.ts serverExternalPackages comment to reflect actual reason
