+++ apps/web/src/test/server-only-stub.ts
@@ -1,0 +1,2 @@
+// Stub for `server-only` in vitest (jsdom has no react-server export condition).
+export default {}
