+++ apps/web/vitest.config.ts
@@ -35,6 +35,7 @@ export default defineConfig({
   resolve: {
     alias: {
       '@': path.resolve(__dirname, './src'),
+      'server-only': path.resolve(__dirname, 'src/test/server-only-stub.ts'),
     },
   },
 })
