+++ package.json
@@ -70,8 +70,7 @@
     "@types/react": "^19.1.13",
     "@types/react-dom": "^19.1.9",
     "mdast-util-to-hast": "^13.2.1",
-    "@react-email/preview-server>next": "16.2.6",
-    "pg": ">=8.16.3 <8.21.0"
+    "@react-email/preview-server>next": "16.2.6"
   },
   "trustedDependencies": [
     "@tensorflow/tfjs-node",
